add support for methods (#4)
This commit adds predicate language support for method calls, such as `set("a").add("b")`.

Realistically these are just a convenience over using bare functions, but functions can be quite onerous in cases where many method calls would be chained. For example, `set("a").add("b").remove("a").intersect(c)` would have to be written as `intersect(remove(add(set("a"), "b"), "a"), c)`, which is hard to read and not very intuitive to write.

No methods are defined be default, they can be registered when creating the parser similarly to how bare functions can be registered today. See methods_test.go for a number of examples. Methods can be implemented by a single type, or multiple types by using an interface.