Create `trace.NotFoundError` once per call to `GetFieldByTag`
Profiles revealed that the majority of time spent in `GetFieldByTag`
was in `trace.NotFound` capturing stack traces to include in the
error. To optimize `GetFieldByTag` a new `notFound` marker error is
returned from the recursive calls and are only converted into a
`trace.NotFoundError` a single time before returning to the original
caller. Since consumers of this library may rely on `trace.IsNotFound`
on the errors returned from `GetFieldByTag` we must still return
a `trace.NotFoundError`.

See https://github.com/gravitational/teleport/issues/27228 for more
details on the memory and cpu consumption caused by `GetFieldByTag`
prior to these changes.